+++ src/application/dataClass/OrderTable.java
@@ -1,10 +1,8 @@
 package application.dataClass;
 
 import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
@@ -16,7 +14,7 @@ public class OrderTable {
 	private StringProperty status;
 	private StringProperty orderDate;
 	private StringProperty comment;
-	private DoubleProperty subTotal;
+	private IntegerProperty subTotal;
 	private StringProperty itemName;
 	private StringProperty Business;
 	// private CheckBoxTCell checkbox;
@@ -59,8 +57,8 @@ public class OrderTable {
 		this.status = new SimpleStringProperty(status);
 	}
 
-	public void setSubTotal(double subTotal) {
-		this.subTotal = new SimpleDoubleProperty(subTotal);
+	public void setSubTotal(int subTotal) {
+		this.subTotal = new SimpleIntegerProperty(subTotal);
 	}
 
 	public String getBusiness() {
@@ -91,7 +89,7 @@ public class OrderTable {
 		return status.get();
 	}
 
-	public double getSubTotal() {
+	public int getSubTotal() {
 		return subTotal.get();
 	}
 }
